ptp2: fix -Wincompatible-pointer-types compiler error in chdk.c
Fixes build error seen with gcc 14.3.0:

ptp2/chdk.c: In function 'yuv_live_to_jpeg':
ptp2/chdk.c:1182:41: error: passing argument 3 of 'jpeg_mem_dest' from incompatible pointer type [-Wincompatible-pointer-types]
 1182 |         jpeg_mem_dest (&cinfo, &outbuf, &outlen);
      |                                         ^~~~~~~
      |                                         |
      |                                         long unsigned int *
In file included from ptp2/chdk.c:32:
/home/autobuild/autobuild/instance-4/output-1/host/mips64el-buildroot-linux-uclibc/sysroot/usr/include/jpeglib.h:979:28: note: expected 'size_t *' {aka 'unsigned int *'} but argument is of type 'long unsigned int *'
  979 | EXTERN(void) jpeg_mem_dest JPP((j_compress_ptr cinfo,

+++ camlibs/ptp2/chdk.c
@@ -1145,7 +1145,7 @@ static void yuv_live_to_jpeg(unsigned char *p_yuv,
 	struct		jpeg_error_mgr jerr;
 	JSAMPROW	row_ptr[1];
 	uint8_t		*outbuf = NULL, *tmprowbuf = NULL;
-	unsigned long	outlen = 0;
+	uint32_t	outlen = 0;
 	unsigned int	row_inc;
 	int		sshift, dshift, xshift, skip;
